backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.3.1
pymongo==4.6.1
pydantic==2.8.2
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
